Cleanup. Moved change_cwd() function to dragonfly_utils from dragonfly_optimizer.
